Work Order Quotation bug fixces

+++ DAL/Models/WorkOrder/WorkOrderQuoteMaterial.cs
@@ -26,7 +26,7 @@ namespace DAL.Models
         public int? BillingMethodId { get; set; }
         public decimal? TMAmount { get; set; }
         public decimal? FlateRate { get; set; }
-        public decimal? HeaderMarkupId { get; set; }
+        public long? HeaderMarkupId { get; set; }
         public int MasterCompanyId { get; set; }
         public string CreatedBy { get; set; }
         public string UpdatedBy { get; set; }
